Run pre-push
